requirements-dev.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
Pillow==8.3.1
python-decouple==3.5
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
typing-extensions==3.10.0.0
